requirements.txt
torch
torchvision
numpy
scipy
opencv-python
imageio
einops
medpy
tqdm
thop
pandas
matplotlib
pyyaml
